finish lab 3 part A

+++ src/kvraft/client.go
@@ -12,7 +12,7 @@ type Clerk struct {
 	// You will have to modify this struct.
 	mu        sync.Mutex
 	clientId  int64
-	requestId int
+	requestId int64
 	leader    int
 }
