+++ src/kvraft/common.go
@@ -16,7 +16,7 @@ type PutAppendArgs struct {
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
 	ClientId  int64
-	RequestId int
+	RequestId int64
 }
 
 type PutAppendReply struct {
@@ -28,7 +28,7 @@ type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
 	ClientId  int64
-	RequestId int
+	RequestId int64
 }
 
 type GetReply struct {
